func/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
nidavellir==0.0.6
pydantic==1.9.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zenpy==2.0.24
